storage: factor away RawStatusUpdate
