+++ src/storage-controller/src/lib.rs
@@ -19,7 +19,6 @@ use std::sync::{Arc, Mutex};
 
 use async_trait::async_trait;
 use chrono::{DateTime, Utc};
-use collection_status::RawStatusUpdate;
 use differential_dataflow::lattice::Lattice;
 use futures::stream::BoxStream;
 use itertools::Itertools;
@@ -1698,7 +1697,7 @@ where
 
         let mut dropped_sources = vec![];
         for id in pending_source_drops.drain(..) {
-            dropped_sources.push(RawStatusUpdate::dropped_status(id, status_now));
+            dropped_sources.push(StatusUpdate::dropped_status(id, status_now));
         }
 
         self.collection_status_manager
@@ -1717,7 +1716,7 @@ where
         {
             let mut sink_statistics = self.sink_statistics.lock().expect("poisoned");
             for id in pending_sink_drops.drain(..) {
-                dropped_sinks.push(RawStatusUpdate::dropped_status(id, status_now));
+                dropped_sinks.push(StatusUpdate::dropped_status(id, status_now));
                 sink_statistics.remove(&id);
             }
         }
@@ -3258,15 +3257,6 @@ where
 
         for update in updates {
             let id = update.id;
-            let update = collection_status::RawStatusUpdate {
-                id,
-                ts: update.timestamp,
-                status_name: update.status,
-                error: update.error,
-                hints: update.hints,
-                namespaced_errors: update.namespaced_errors,
-            };
-
             if self.exports.contains_key(&id) {
                 sink_status_updates.push(update);
             } else if self.collections.contains_key(&id) {
